requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
google-generativeai==0.3.0
pandas==2.1.1
python-dotenv==1.0.0
